+++ IncohAutoCorrelate/ACMesh.cpp
@@ -218,7 +218,7 @@ void ACMesh::Atomic_Add_q_Entry(float q_local[3], float RotationM[9], float Valu
 		return;
 	}
 
-	ArrayOperators::MultiplyScalar(q_local, (double)Shape.Voxel_per_dq, 3);
+	ArrayOperators::MultiplyScalar(q_local, (float)Shape.Voxel_per_dq, 3);
 
 	if (DoubleBinning)
 	{
@@ -383,3 +383,4 @@ void ACMesh::Atomic_Add_rotated_q_Entry(float q_local[3], float Value, Settings:
 		Mesh[(fs + 1) + (ms + 1) * s + (ss + 1) * s*s] += val;
 	}
 }
+
